+++ frontend/app/routes/researcher/metadata.js
@@ -1,0 +1,9 @@
+import Ember from 'ember';
+
+export default Ember.Route.extend({
+
+
+model() {
+  return this.modelFor('researcher');
+}
+});
