+++ frontend/app/templates/components/researcher-form.hbs
@@ -1,21 +1,1 @@
 {{yield}}
-<div class="form-table">
-
-
-  <div class="dropdownHolder">
-    <select class="departmentSelect select" onchange={{action (mut departmentId) value="target.value"}}>
-        <option>Select a Department</option>
-      {{#each model.departments as |departmentChoice|}}
-        <option selected={{eq department departmentChoice}} value={{departmentChoice.id}}>{{departmentChoice.name}}</option>
-      {{/each}}
-    </select>
-    {{input value=model.grant.number class="form-control" id="grantid" placeholder="Grant Number"}}
-  </div>
-
-  <button class="btn btn-success addFunding" {{action "submit"}}>Add More Funding</button>
-
-   {{file-chooser files=fileList}}
-
-   <button class="btn btn-primary btn-lg btn-block metaData"  {{action 'buttonClicked' model.grant departmentId fileList model.document}} > Next </button>
-
-</div>
